Updated to include authentication & authorization

+++ server/customers/customers.controller.ts
@@ -1,5 +1,6 @@
 import * as bodyParser from "body-parser";
 import * as express from "express";
+import Token from "../util/auth";
 import Customer from "./customers.model";
 
 const router = express.Router();
@@ -10,8 +11,7 @@ router.route("/").post(bodyParser.json(), async (request, response) => {
         await customer.save();
         return response.status(201).json(customer);
     } catch (error) {
-        console.log(error);
-        return response.status(400).send(error);
+        return response.status(400).json(error.toString());
     }
 });
 
@@ -26,42 +26,50 @@ router.route("/:id/exist").get(async (request, response) => {
             throw new Error("Not Found");
         }
     } catch (error) {
-        console.log(error);
-        return response.status(404).send(false);
+        return response.status(404).json(false);
     }
 });
 
-router.route("/:id").get(async (request, response) => {
+router.route("/:id").get(Token.authenticate, async (request, response) => {
     try {
-        const customerId = request.params.id;
-        const customer = await Customer.find({ _id: customerId });
-        return response.status(200).json(customer);
+        if (await Token.authorize(["customer"], request)) {
+            const customerId = request.params.id;
+            const customer = await Customer.find({ _id: customerId });
+            return response.status(200).json(customer);
+        } else {
+            throw Error("No Access");
+        }
     } catch (error) {
-        console.log(error);
-        return response.status(404).send(error);
+        return response.status(404).json(error.toString());
     }
 });
 
-router.route("/:id").delete(async (request, response) => {
+router.route("/:id").delete(Token.authenticate, async (request, response) => {
     try {
-        const customerId = request.params.id;
-        await Customer.findOneAndRemove({ _id: customerId });
-        return response.status(202).json("Customer deleted!");
+        if (Token.authorize(["owner"], request)) {
+            const customerId = request.params.id;
+            await Customer.findOneAndRemove({ _id: customerId });
+            return response.status(202).json("Customer deleted!");
+        } else {
+            throw Error("No Access");
+        }
     } catch (error) {
-        console.log(error);
-        return response.status(404).send(error);
+        return response.status(404).json(error.toString());
     }
 });
 
-router.route("/:id").put(bodyParser.json(), async (request, response) => {
+router.route("/:id").put(Token.authenticate, bodyParser.json(), async (request, response) => {
     try {
-        const customerId = request.params.id;
-        const customerUpdate = request.body;
-        await Customer.update({ _id: customerId }, customerUpdate, {new: true});
-        return response.status(202).json("Customer updated!");
+        if (Token.authorize(["customer"], request)) {
+            const customerId = request.params.id;
+            const customerUpdate = request.body;
+            await Customer.update({ _id: customerId }, customerUpdate, { new: true });
+            return response.status(202).json("Customer updated!");
+        } else {
+            throw Error("No Access");
+        }
     } catch (error) {
-        console.log(error);
-        return response.status(404).send(error);
+        return response.status(404).json(error.toString());
     }
 });
